test(contests): add unit tests for scoring calculations and ranking generation
